# backend/openapi.yml
openapi: 3.0.3
info:
  title: STGY API
  version: 1.0.0
  description: |
    STGY SNS backend API (Express/TypeScript)
    - Cookie authentication: session_id
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    NotificationRecordUser:
      type: object
      properties:
        userId: { type: string }
        userNickname: { type: string }
        ts: { type: number }
      required: [userId, userNickname, ts]
    NotificationRecordPost:
      type: object
      properties:
        userId: { type: string }
        userNickname: { type: string }
        postId: { type: string }
        postSnippet: { type: string }
        ts: { type: number }
      required: [userId, userNickname, postId, postSnippet, ts]
    Notification:
      type: object
      properties:
        slot: { type: string }
        term: { type: string }
        isRead: { type: boolean }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        countUsers: { type: integer }
        countPosts: { type: integer }
        records:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/NotificationRecordUser"
              - $ref: "#/components/schemas/NotificationRecordPost"
      required: [slot, term, isRead, updatedAt, createdAt, records]
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /notifications/feed:
    get:
      summary: Get notification feed
      security:
        - sessionCookie: []
      parameters:
        - name: newerThan
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Notification feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "304":
          description: Not Modified (no updates since newerThan)
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "400":
          $ref: "#/components/responses/Error400"
  /notifications/mark:
    post:
      summary: Mark a single notification (by slot + term)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slot: { type: string }
                term: { type: string }
                isRead: { type: boolean }
              required: [slot, term, isRead]
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /notifications/mark-all:
    post:
      summary: Mark all notifications for current user
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isRead: { type: boolean }
              required: [isRead]
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
...